Remove roles field from org member resource
